Remove unhelpful code
The read() method runs on each host, where there is no configuration or
knowledge of the maximum_use_percentage.  Hosts should not report this
to the server, as they will always fall back to the default value.

The alarm() method runs on the server, so it should consult the database
for a per-host maximum_use_percentage value, falling back to the default
if necessary.

This fix has the convenient side effect of making the code simpler.