Improve uefulness of CPU agent
The CPU agent was previously reading the average load across all
processes, which is not very useful for monitoring high load on a
server.

This commit changes to a snapshot-in-time of the average CPU load across
all CPUs. It also adds an alarm of 50%, which may need to be increased.